# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Cottage Bakery API
  version: "0.1.0"
  description: |
    The HTTP contract between `web/` (Next.js, the only client of this
    API) and `api/` (Express, the only thing that talks to Supabase).

    This file is the single source of truth for that contract — both
    services generate TypeScript types from it (`npm run gen:types` in
    each) rather than hand-writing request/response types twice. See
    ../specs/ENGINEERING_RULES.md and ../docs/adr/0004-service-boundary.md
    for why this exists at all: without a shared npm package between two
    independently-deployable services (a deliberate choice — see the
    ADR), this spec is what keeps the boundary honest instead of the two
    sides quietly drifting apart.

    Nothing in this file is reachable from a browser. `web/`'s own API
    routes and Server Components call this API server-side and are the
    only callers — see docs/ARCHITECTURE.md. One exception exists outside
    this file's scope entirely: `POST /webhooks/paypal` is called by
    PayPal's own servers, not by `web/`, so it's deliberately left out of
    this web/-api/ contract — see
    docs/adr/0006-online-payment-paypal.md and api/src/routes/webhooks.ts.
servers:
  - url: http://localhost:4000
    description: Local dev
paths:
  /health:
    get:
      operationId: health
      summary: Liveness check.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean

  /config:
    get:
      operationId: getConfig
      summary: |
        Non-secret runtime config `web/` needs but doesn't hold itself —
        currently just PayPal's Client ID (see
        docs/adr/0006-online-payment-paypal.md for why it lives here
        instead of being duplicated into web/'s own env).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [paypalClientId]
                properties:
                  paypalClientId:
                    type: [string, "null"]
                    description: Null if PAYPAL_CLIENT_ID isn't configured in this environment.

  /products:
    get:
      operationId: listProducts
      summary: Active menu items, ordered by category then name.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      summary: |
        Create a product. The slug comes from the name (made unique with a
        numeric suffix if needed). Starts hidden unless isActive is true.
        Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/Errors"
        "401":
          $ref: "#/components/responses/Errors"

  /products/{id}:
    patch:
      operationId: updateProduct
      summary: Show (isActive true) or hide a product on the menu. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/Errors"
        "401":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /products/all:
    get:
      operationId: listAllProducts
      summary: Every product, including hidden ones (isActive false). Admin only.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Errors"

  /products/{id}/image:
    put:
      operationId: setProductImage
      summary: |
        Upload a product photo, replacing any existing one. The request body
        is the image itself (at most 4 MB). Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"
        "413":
          $ref: "#/components/responses/Errors"
        "415":
          $ref: "#/components/responses/Errors"
    delete:
      operationId: deleteProductImage
      summary: Remove a product's photo. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /products/{id}/sold-out:
    patch:
      operationId: setProductSoldOut
      summary: |
        Mark a product sold out for today (bakery date, America/Los_Angeles),
        or clear it. Resets on its own at midnight; orders for later dates
        are unaffected. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [soldOutToday]
              properties:
                soldOutToday:
                  type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [product]
                properties:
                  product:
                    $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/Errors"
        "401":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /orders:
    post:
      operationId: createOrder
      summary: |
        Submit an order request. Prices are looked up server-side from
        `productId`s; the client-supplied price, if any, is ignored. An
        order exists in `PENDING` / `UNPAID` regardless of
        `paymentMethod` — for a `PAYPAL` order, payment happens in a
        separate step after this call (see `POST /orders/{id}/paypal-order`
        and `POST /orders/{id}/capture-payment`), not as part of creating
        the order. See docs/adr/0002-order-fulfillment-model.md and
        docs/adr/0006-online-payment-paypal.md.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created.
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/OrderWithItems"
        "400":
          $ref: "#/components/responses/Errors"
    get:
      operationId: listOrders
      summary: All orders, newest first. Admin only.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderWithItems"
        "401":
          $ref: "#/components/responses/Errors"

  /orders/{id}:
    get:
      operationId: getOrder
      summary: |
        A single order, by id. Public — the order id itself (a uuid) is
        the access control, same as the confirmation-page link a
        customer gets after ordering. No customer accounts exist to
        authenticate against.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/OrderWithItems"
        "404":
          $ref: "#/components/responses/Errors"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      summary: Change an order's status. Admin only.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/Errors"
        "401":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /orders/{id}/paypal-order:
    post:
      operationId: createPayPalOrderForOrder
      summary: |
        Starts a PayPal Checkout for an existing order — creates a
        PayPal order sized to this order's own `totalCents` (never a
        client-supplied amount) and returns PayPal's order id for the
        PayPal JS SDK to use client-side. See
        docs/adr/0006-online-payment-paypal.md.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [paypalOrderId]
                properties:
                  paypalOrderId:
                    type: string
        "400":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /orders/{id}/capture-payment:
    post:
      operationId: capturePaymentForOrder
      summary: |
        Captures a buyer-approved PayPal payment for this order,
        server-side — never trusts a client-reported "payment succeeded."
        On success, the order's `paymentStatus` becomes `PAID` and its
        `status` becomes `CONFIRMED`. Idempotent: calling this again for
        an order that's already `PAID` returns the same success response
        rather than erroring. See docs/adr/0006-online-payment-paypal.md.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paypalOrderId]
              properties:
                paypalOrderId:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/Errors"
        "404":
          $ref: "#/components/responses/Errors"

  /admin/login:
    post:
      operationId: adminLogin
      summary: |
        Exchange the shared admin password for a bearer token. This
        service issues the token but never sets a cookie — it has no
        concept of a browser, only of a bearer token. `web/` is the one
        that turns this token into an httpOnly cookie (see
        web/src/app/api/admin/login/route.ts).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
        "401":
          $ref: "#/components/responses/Errors"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Issued by POST /admin/login. web/ forwards this as
        `Authorization: Bearer <token>` on every admin request; it never
        verifies the token's signature itself — only this service holds
        JWT_SECRET. See docs/adr/0004-service-boundary.md.
  parameters:
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Errors:
      description: |
        This API's one error envelope shape, used for every non-2xx
        response — see specs/ENGINEERING_RULES.md "Error envelope."
        Never `{ "error": "..." }` (singular) — that inconsistency
        existed in the pre-split codebase and was deliberately fixed
        when this spec was written, not carried forward.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            type: string

    FulfillmentMethod:
      type: string
      enum: [PICKUP, LOCAL_DELIVERY, IN_STATE_SHIPPING]

    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, READY, COMPLETED, CANCELLED]

    PaymentMethod:
      type: string
      enum: [MANUAL, PAYPAL]
      description: |
        MANUAL = cash/Venmo/Zelle, settled directly with the baker at
        pickup/delivery (the original, and still-default, model). PAYPAL
        = paid online through PayPal Checkout. See
        docs/adr/0006-online-payment-paypal.md.

    PaymentStatus:
      type: string
      enum: [UNPAID, PAID]

    Product:
      type: object
      required:
        [id, slug, name, description, priceCents, category, imageUrl, allergens, ingredients, netWeight, isActive, soldOutOn, soldOutToday, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        priceCents:
          type: integer
          description: Always integer cents — see specs/ENGINEERING_RULES.md "Money."
        category:
          type: string
        imageUrl:
          type: [string, "null"]
        allergens:
          type: string
        ingredients:
          type: string
          description: Full ingredients statement, as required on a California cottage food product label.
        netWeight:
          type: string
          description: Net weight as printed on the label, e.g. "8 oz (226 g)" — a formatted display string, not a numeric quantity to compute with.
        isActive:
          type: boolean
        soldOutOn:
          type: [string, "null"]
          format: date
          description: Bakery date the admin marked this sold out, or null.
        soldOutToday:
          type: boolean
          description: True when soldOutOn is today's bakery date. Blocks orders for today only.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateProductRequest:
      type: object
      required: [name, description, priceCents, category, allergens, ingredients, netWeight]
      properties:
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
        priceCents:
          type: integer
          minimum: 1
          maximum: 100000
        category:
          type: string
          maxLength: 40
        allergens:
          type: string
          maxLength: 200
          description: Comma-separated major food allergens, e.g. "wheat, milk, eggs", or "none".
        ingredients:
          type: string
          maxLength: 2000
        netWeight:
          type: string
          maxLength: 60
          description: As printed on the label, e.g. "12 oz (340 g)".
        isActive:
          type: boolean
          default: false

    OrderItemWithProduct:
      type: object
      required: [id, orderId, productId, quantity, unitPriceCents, product]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
        unitPriceCents:
          type: integer
          description: Snapshot of the product's price at order time — not a live reference. See supabase/migrations/0001_init.sql.
        product:
          $ref: "#/components/schemas/Product"

    Order:
      type: object
      required:
        [id, customerName, customerEmail, customerPhone, fulfillmentMethod, fulfillmentAddress, requestedDate, notes, status, subtotalCents, deliveryFeeCents, totalCents, paymentMethod, paymentStatus, paypalOrderId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        fulfillmentMethod:
          $ref: "#/components/schemas/FulfillmentMethod"
        fulfillmentAddress:
          type: [string, "null"]
        requestedDate:
          type: string
          description: |
            A bare "YYYY-MM-DD" date-only string, not a timestamp — see
            specs/ENGINEERING_RULES.md "Date-only values." Never parse
            this with `new Date(str)` on either side of the boundary.
        notes:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/OrderStatus"
        subtotalCents:
          type: integer
        deliveryFeeCents:
          type: integer
          description: Computed server-side at order creation (never taken from the request). 0 for pickup, shipping, and deliveries at or above the free-delivery threshold. See docs/adr/0008-delivery-fee.md.
        totalCents:
          type: integer
          description: subtotalCents + deliveryFeeCents. What the customer owes, and what a PayPal order is created for.
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paypalOrderId:
          type: [string, "null"]
          description: Set once a PayPal Checkout has been started for this order (see POST /orders/{id}/paypal-order). Null for a MANUAL order, or a PAYPAL order for which checkout hasn't started yet.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItemWithProduct"

    CreateOrderRequest:
      type: object
      required: [customerName, customerEmail, customerPhone, fulfillmentMethod, requestedDate, items]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        fulfillmentMethod:
          $ref: "#/components/schemas/FulfillmentMethod"
        fulfillmentAddress:
          type: string
        requestedDate:
          type: string
          description: "YYYY-MM-DD — see Order.requestedDate."
        notes:
          type: string
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
          description: Omit for the default (MANUAL).
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1